Move CupsPrinterFFM to oshi.hardware.platform.unix (mirror JNA peer)
`CupsPrinterFFM` is an implementation class (extends the abstract
`CupsPrinter` from `oshi-common`), not an FFM mapping. Its JNA peer
`CupsPrinterJNA` lives at `oshi.hardware.platform.unix.CupsPrinterJNA`;
this move lines the two up.

It was sharing `oshi.ffm.unix` with `CupsFunctions` (the actual libcups
FFM mapping, which stays put). The same-package access to
`CupsFunctions` becomes an explicit import.

Six consumers (the `*HardwareAbstractionLayerFFM` classes on Linux,
Mac, FreeBSD, OpenBSD, Solaris, and AIX) had their imports rewritten.
